discord-bot-python/pyproject.toml
[project]
name = "foxhole-quartermaster-bot"
version = "1.0.0"
description = "Discord bot for Foxhole Quartermaster - AI-powered logistics assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
